Add new management layer not using mongodb
- Locations and accounts are read from standard configuration files
- Workflows are passively applied through an API call
- Workflow state is stored in Zookeeper
