+++ lib/api/BackbeatAPI.js
@@ -30,6 +30,10 @@ const {
     zkStatePath: zkIngestionStatePath,
 } = require('../../extensions/ingestion/constants');
 
+const { applyBucketReplicationWorkflows } = require('../../extensions/replication/management');
+const { applyBucketLifecycleWorkflows } = require('../../extensions/lifecycle/management');
+const { createServiceState } = require('../../lib/management');
+
 // StatsClient constant defaults
 // TODO: This should be moved to constants file
 const INTERVAL = 300; // 5 minutes
@@ -53,6 +57,7 @@ class BackbeatAPI {
         this._kafkaConfig = config.kafka;
         this._repConfig = config.extensions.replication;
         this._ingestionConfig = config.extensions.ingestion;
+        this._lifecycleConfig = config.extensions.lifecycle;
         this._queuePopulator = config.queuePopulator;
         this._kafkaHost = config.kafka.hosts;
         this._redisConfig = config.redis;
@@ -211,6 +216,16 @@ class BackbeatAPI {
                 r.category === 'monitoring')[0];
             bbRequest.setMatchedRoute(promRoute);
             bbRequest.setContentType(ZenkoMetrics.asPrometheusContentType());
+            return null;
+        }
+        if (route === 'configuration/workflows') {
+            if (bbRequest.getHTTPMethod() !== 'POST') {
+                return errors.MethodNotAllowed;
+            }
+            const configRoute = routes.filter(r =>
+                r.category === 'configuration')[0];
+            bbRequest.setMatchedRoute(configRoute);
+            bbRequest.setContentType('application/json');
             return null;
         }
 
@@ -1285,6 +1300,56 @@ class BackbeatAPI {
             });
             return cb(null, statuses);
         });
+    }
+
+    /**
+     * Apply Workflow configuration
+     * @param {Object} details - The route details
+     * @param {String} body - The POST request body string
+     * @param {Function} cb - The callback to call
+     * @return {undefined}
+     */
+    applyWorkflowConfiguration(details, body, cb) {
+        let overlay;
+        try {
+            overlay = JSON.parse(body);
+        } catch (e) {
+            const err = errors.MalformedPOSTRequest.customizeDescription(
+                'invalid JSON');
+            return cb(err);
+        }
+        this._logger.debug('applying, parsed json', { overlay });
+
+        const serviceStateJobs = [
+            {
+                serviceName: 'replication',
+                serviceAccount: this._repConfig.source.auth.account,
+                applyBucketWorkflows: applyBucketReplicationWorkflows,
+            },
+            {
+                serviceName: 'lifecycle',
+                serviceAccount: this._lifecycleConfig.auth.account,
+                applyBucketWorkflows: applyBucketLifecycleWorkflows,
+            },
+        ];
+
+        return async.each(
+            serviceStateJobs,
+            (jobParams, done) => {
+                this._logger.info('applying', { jobParams });
+                const serviceState = createServiceState(jobParams.serviceName, this._zkClient);
+                serviceState.apply(overlay, jobParams, done);
+            },
+            err => {
+                if (err) {
+                    this._logger.error('could not apply service state', {
+                        error: err,
+                    });
+                    return cb(errors.InternalError, {});
+                }
+                return cb(null, {});
+            }
+        );
     }
 
     /**
