+++ lib/api/routes.js
@@ -202,6 +202,14 @@ function routes(locations) {
             },
             method: 'getServiceStatus',
         },
+        // Route: /_/configuration/workflows
+        {
+            httpMethod: 'POST',
+            category: 'configuration',
+            type: 'workflows',
+            method: 'applyWorkflowConfiguration',
+            extensions: {},
+        },
     ];
 }
