+++ lib/management/patchConfiguration.js
@@ -77,4 +77,7 @@ logger, cb) {
     return process.nextTick(cb, null);
 }
 
-module.exports = patchConfiguration;
+module.exports = {
+    patchConfiguration,
+    updateLocations,
+};
